move hvplot_extension to the utilities module
